:recycle: Refactor: Member 도메인 Facade계층 추가
