+++ src/main/java/briefing/briefing/domain/Briefing.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 
 import jakarta.persistence.*;
 
-import briefing.base.BaseDateTimeEntity;
+import briefing.common.domain.BaseDateTimeEntity;
 import briefing.common.enums.BriefingType;
 import briefing.common.enums.GptModel;
 import briefing.common.enums.TimeOfDay;
