+++ src/main/java/briefing/briefing/domain/BriefingArticle.java
@@ -8,7 +8,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
-import briefing.base.BaseDateTimeEntity;
+import briefing.common.domain.BaseDateTimeEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
