+++ src/main/java/briefing/member/implement/MemberQueryService.java
@@ -1,18 +1,19 @@
 package briefing.member.implement;
 
+import java.util.Optional;
+
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import briefing.common.enums.MemberRole;
 import briefing.common.enums.SocialType;
 import briefing.common.exception.common.ErrorCode;
+import briefing.infra.redis.domain.RefreshToken;
 import briefing.member.domain.Member;
 import briefing.member.domain.repository.MemberRepository;
 import briefing.member.exception.MemberException;
 import lombok.RequiredArgsConstructor;
 
 @Service
-@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class MemberQueryService {
     private final MemberRepository memberRepository;
@@ -23,7 +24,10 @@ public class MemberQueryService {
                 .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
     }
 
-    @Transactional
+    public Optional<Member> findBySocialIdAndSocialType(String socialId, SocialType socialType) {
+        return memberRepository.findBySocialIdAndSocialType(socialId, socialType);
+    }
+
     public Member testForTokenApi() {
         return memberRepository
                 .findFirstByOrderByCreatedAt()
@@ -36,5 +40,11 @@ public class MemberQueryService {
                                                 .socialType(SocialType.GOOGLE)
                                                 .role(MemberRole.ROLE_USER)
                                                 .build()));
+    }
+
+    public Member parseRefreshToken(RefreshToken refreshToken) {
+        return memberRepository
+                .findById(refreshToken.getMemberId())
+                .orElseThrow(() -> new MemberException(ErrorCode.MEMBER_NOT_FOUND));
     }
 }
