+++ src/main/java/briefing/member/presentation/MemberV2Api.java
@@ -1,26 +1,17 @@
 package briefing.member.presentation;
 
-import java.util.List;
-
 import jakarta.validation.Valid;
 
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import briefing.common.enums.MemberRole;
 import briefing.common.enums.SocialType;
 import briefing.common.presentation.response.CommonResponse;
-import briefing.infra.redis.domain.RefreshToken;
-import briefing.infra.redis.service.RedisService;
-import briefing.member.business.MemberConverter;
+import briefing.member.business.MemberFacade;
 import briefing.member.domain.Member;
-import briefing.member.implement.MemberCommandService;
-import briefing.member.implement.MemberQueryService;
 import briefing.member.presentation.dto.MemberRequest;
 import briefing.member.presentation.dto.MemberResponse;
 import briefing.security.handler.annotation.AuthMember;
-import briefing.security.provider.TokenProvider;
 import briefing.validation.annotation.CheckSameMember;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -44,12 +35,7 @@ import lombok.RequiredArgsConstructor;
 })
 @RequestMapping("/v2")
 public class MemberV2Api {
-    private final MemberQueryService memberQueryService;
-    private final MemberCommandService memberCommandService;
-
-    private final TokenProvider tokenProvider;
-
-    private final RedisService redisService;
+    private final MemberFacade memberFacade;
 
     @Operation(summary = "02-01 Member\uD83D\uDC64 소셜 로그인 V2", description = "구글, 애플 소셜로그인 API입니다.")
     @PostMapping("/members/auth/{socialType}")
@@ -57,19 +43,7 @@ public class MemberV2Api {
             @Parameter(description = "소셜로그인 종류", example = "google") @PathVariable
                     final SocialType socialType,
             @RequestBody final MemberRequest.LoginDTO request) {
-        Member member = memberCommandService.login(socialType, request);
-        String accessToken =
-                tokenProvider.createAccessToken(
-                        member.getId(),
-                        member.getSocialType().toString(),
-                        member.getSocialId(),
-                        List.of(new SimpleGrantedAuthority(MemberRole.ROLE_USER.name())));
-        String refreshToken =
-                redisService
-                        .generateRefreshToken(member.getSocialId(), member.getSocialType())
-                        .getToken();
-        return CommonResponse.onSuccess(
-                MemberConverter.toLoginDTO(member, accessToken, refreshToken));
+        return CommonResponse.onSuccess(memberFacade.login(socialType, request));
     }
 
     @Operation(
@@ -93,17 +67,7 @@ public class MemberV2Api {
     @PostMapping("/members/auth/token")
     public CommonResponse<MemberResponse.ReIssueTokenDTO> reissueTokenV2(
             @Valid @RequestBody MemberRequest.ReissueDTO request) {
-        RefreshToken refreshToken = redisService.reGenerateRefreshToken(request);
-        Member parsedMember = memberCommandService.parseRefreshToken(refreshToken);
-        String accessToken =
-                tokenProvider.createAccessToken(
-                        parsedMember.getId(),
-                        parsedMember.getSocialType().toString(),
-                        parsedMember.getSocialId(),
-                        List.of(new SimpleGrantedAuthority(parsedMember.getRole().toString())));
-        return CommonResponse.onSuccess(
-                MemberConverter.toReIssueTokenDTO(
-                        parsedMember.getId(), accessToken, refreshToken.getToken()));
+        return CommonResponse.onSuccess(memberFacade.reIssueToken(request));
     }
 
     @Operation(summary = "02-01 Member\uD83D\uDC64 회원 탈퇴 V2", description = "회원 탈퇴 API 입니다.")
@@ -137,7 +101,6 @@ public class MemberV2Api {
     })
     public CommonResponse<MemberResponse.QuitDTO> quitMemberV2(
             @AuthMember Member member, @CheckSameMember @PathVariable Long memberId) {
-        memberCommandService.deleteMember(memberId);
-        return CommonResponse.onSuccess(MemberConverter.toQuitDTO());
+        return CommonResponse.onSuccess(memberFacade.quit(memberId));
     }
 }
